Fix: ConcurrentModificationException caused by multiple status command registrations on player join; command now registers only once on plugin startup
