+++ src/main/kotlin/yv/tils/smp/mods/status/StatusJoinQuit.kt
@@ -3,6 +3,7 @@ package yv.tils.smp.mods.status
 import org.bukkit.event.player.PlayerJoinEvent
 import org.bukkit.event.player.PlayerQuitEvent
 import yv.tils.smp.utils.color.ColorUtils
+import yv.tils.smp.utils.configs.global.Config
 import yv.tils.smp.utils.configs.language.LangStrings
 import yv.tils.smp.utils.configs.language.Language
 import yv.tils.smp.utils.configs.status.StatusConfig
@@ -10,11 +11,15 @@ import yv.tils.smp.utils.internalAPI.Placeholder
 
 class StatusJoinQuit {
     fun loadPlayer(e: PlayerJoinEvent) {
+        if (!(Config.config["modules.status"] as Boolean)) {
+            return
+        }
+
         val player = e.player
         val status = StatusConfig.saves[player.uniqueId.toString()]
 
         if (status != null) {
-            if (StatusCommand().setStatusDisplay(player, status as String)) {
+            if (StatusHandler().setStatusDisplay(player, status as String)) {
                 val display = StatusConfig.config["display"] as String
 
                 val displayCompo = Placeholder().replacer(
